RESOLVE: delete `PatResolver`
Part of #207
